Removing unnecesarry abstract class which checks for token expiry - It's probably a better idea to always check for token expiry as a default behaviour
